requirements-dev.txt
ultralytics
pandas
pillow-heif
pillow
scikit-learn
numpy
streamlit
pyyaml
